feat : agrege lo de filtar admes ajuste todas las otras funciones q me daban error me flata lo de la img url de los cohes y las 2 bonificaciones esas

+++ app/src/main/java/com/pmd/rentavehiculos/model/RentaRequest.kt
@@ -4,7 +4,7 @@ data class RentaRequest(
     val persona: Persona,  // ✅ Ahora tendrá `nombre` y `apellidos`
     val vehiculo: Vehiculo,
     val dias_renta: Int,
-    val valor_total_renta: Double,
+    val valor_total_renta: Int,
     val fecha_renta: String,
     val fecha_estimada_entrega: String
 )
